Add android version of Example
