+++ Example/android/app/build.gradle
@@ -133,6 +133,7 @@ android {
 }
 
 dependencies {
+    compile project(':react-native-shimmer')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
